api: Add a function to configure depth processors
Since direct access to depth processors is removed, add
Freenect2Device::setConfiguration() to allow users to
configure depth processors. This design is consistent with
IrCameraParams also being processed in Freenect2Device.
