+++ include/libfreenect2/libfreenect2.hpp
@@ -82,6 +82,18 @@ public:
     float fx, fy, cx, cy, k1, k2, k3, p1, p2;
   };
 
+  /** Configuration of depth processing. */
+  struct Config
+  {
+    float MinDepth;
+    float MaxDepth;
+
+    bool EnableBilateralFilter; ///< Whether to run the bilateral filter.
+    bool EnableEdgeAwareFilter; ///< Whether to run the edge aware filter.
+
+    Config();
+  };
+
   virtual ~Freenect2Device();
 
   virtual std::string getSerialNumber() = 0;
@@ -91,6 +103,7 @@ public:
   virtual Freenect2Device::IrCameraParams getIrCameraParams() = 0;
   virtual void setColorCameraParams(const Freenect2Device::ColorCameraParams &params) = 0;
   virtual void setIrCameraParams(const Freenect2Device::IrCameraParams &params) = 0;
+  virtual void setConfiguration(const Config &config) = 0;
 
   virtual void setColorFrameListener(libfreenect2::FrameListener* rgb_frame_listener) = 0;
   virtual void setIrAndDepthFrameListener(libfreenect2::FrameListener* ir_frame_listener) = 0;
